Cut the silent minute before the first log line on startup
The app spent ~11 s of CPU importing aiogram, pyrogram, adaptix, matplotlib
and friends before `setup_logging` was ever reached, because every entrypoint
pulled the whole application in at module level and only configured logging
inside `main()`. Under the 0.15 CPU quota the backend runs with, that is over
a minute of a process that looks completely hung.

Two independent causes, both addressed:

* Nothing was logged until the imports were done. Every entrypoint is now a
  thin bootstrap that configures logging first and only then imports the
  application, reporting how long that import took. The combined app factory
  moved from `shvatka/__main__.py` to `shvatka.main_factory.create_root_app`.
  Time from process start to the first log line drops from ~11 s to ~0.3 s
  of CPU.

* The image shipped no bytecode, so every fresh container compiled the whole
  venv on first import. Building with UV_COMPILE_BYTECODE=1 and running
  compileall over the app package cuts the import itself from ~11 s to ~4.8 s
  of CPU.
