+++ shvatka/main_factory.py
@@ -1,9 +1,23 @@
 import logging
+from functools import partial
 
-from dishka import Provider, Scope, AsyncContainer, provide
+from aiogram import Bot, Dispatcher
+from dishka import (
+    Provider,
+    Scope,
+    AsyncContainer,
+    provide,
+    make_async_container,
+    plotter,
+    STRICT_VALIDATION,
+)
 from dishka.exceptions import NoContextValueError
+from dishka.integrations.fastapi import setup_dishka
+from fastapi import FastAPI
 
+from shvatka.api.app.dependencies import get_api_specific_providers
 from shvatka.api.app.dependencies.auth import ApiIdentityProvider
+from shvatka.api.app.config.parser.main import load_config as load_api_config
 from shvatka.api.app.utils.web_input import (
     WebGameView,
     WebTeamNotifier,
@@ -11,6 +25,8 @@ from shvatka.api.app.utils.web_input import (
     WebGamePreparer,
     WebGameLogWriter,
 )
+from shvatka.api.main_factory import create_app
+from shvatka.common.config.models.paths import Paths
 from shvatka.core.interfaces.identity import IdentityProvider
 from shvatka.core.views.game import (
     GameView,
@@ -19,7 +35,16 @@ from shvatka.core.views.game import (
     GameLogWriter,
 )
 from shvatka.core.views.team import TeamNotifier
+from shvatka.infrastructure.di import get_providers
+from shvatka.infrastructure.di.utils import warm_up
+from shvatka.tgbot.config.models.bot import WebhookConfig
+from shvatka.tgbot.config.parser.main import load_config as load_bot_config
+from shvatka.tgbot.main_factory import (
+    resolve_update_types,
+    get_bot_specific_providers,
+)
 from shvatka.tgbot.services.identity import TgBotIdentityProvider
+from shvatka.tgbot.utils.fastapi_webhook import setup_application, SimpleRequestHandler
 from shvatka.tgbot.views.game import BotView, BotOrgNotifier, GameBotLog
 from shvatka.tgbot.views.team import BotTeamNotifier
 from shvatka.views import (
@@ -70,3 +95,50 @@ def get_complex_only_providers() -> list[Provider]:
     return [
         ComplexOnlyProvider(),
     ]
+
+
+def create_root_app(paths: Paths) -> FastAPI:
+    api_config = load_api_config(paths)
+    bot_config = load_bot_config(paths)
+    webhook_config = bot_config.bot.webhook
+    if not webhook_config:
+        raise EnvironmentError("No webhook configuration provided")
+
+    app = create_app(api_config)
+    dishka = make_async_container(
+        *get_providers("SHVATKA_PATH"),
+        *get_bot_specific_providers(),
+        *get_api_specific_providers(),
+        *get_complex_only_providers(),
+        validation_settings=STRICT_VALIDATION,
+    )
+    setup_application(app, dishka)
+    webhook_handler = SimpleRequestHandler(
+        handle_in_background=False,
+        secret_token=webhook_config.secret,
+    )
+    webhook_handler.register(app, webhook_config.path)
+
+    root_app = FastAPI()
+    root_app.mount(api_config.context_path, app)
+    setup = partial(on_startup, dishka, webhook_config)
+    root_app.router.add_event_handler("startup", setup)
+    setup_dishka(dishka, root_app)
+    logger.info(
+        "app prepared with dishka:\n%s",
+        plotter.render_d2(dishka),
+    )
+    return root_app
+
+
+async def on_startup(dishka: AsyncContainer, webhook_config: WebhookConfig):
+    webhook_url = webhook_config.web_url + webhook_config.path
+    logger.info("as webhook url used %s", webhook_url)
+    bot = await dishka.get(Bot)
+    dp = await dishka.get(Dispatcher)
+    await bot.set_webhook(
+        url=webhook_url,
+        secret_token=webhook_config.secret,
+        allowed_updates=resolve_update_types(dp),
+    )
+    await warm_up(dishka)
